+++ changelog.md
@@ -4,6 +4,7 @@ TGUI 1.11  (TBD)
 - Added RoundedBorderRadius property to EditBox, ComboBox and Tabs renderers
 - Added getLeftPlusRight and getTopPlusBottom functions to Outline class
 - Added `TGUI_BUILD_CXX20_MODULE` option to CMake to build TGUI as a c++20 module
+- ScrollablePanel didn't transform the right click position for its child widgets
 
 
 TGUI 1.10  (14 June 2025)